requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-openai
langchain-community
langchain-pinecone
pinecone
openai
tiktoken
pinecone-client
pypdf
beautifulsoup4
lxml
tqdm